requirements.txt
python-docx==0.8.10
pandas==1.2.3
Pillow==8.1.2
Flask==1.1.2
requests==2.25.1